+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 10.9
 - Replace "Course Title" in TipMessage with actual Course title in Attendance/, Eligibility/ & Grades/TeacherCompletion.php
 - Handle multiple Course Periods on the same Period for same Teacher in Attendance/ & Eligibility/TeacherCompletion.php
 - JS Remove "Go" button & submit form on select change in Eligibility/TeacherCompletion.php
+- Add SetUserCoursePeriod() function in Current.php
 
 Changes in 10.8.4
 -----------------